cal 4

+++ lib/helpers/db_helper.dart
@@ -86,7 +86,7 @@ class DB {
         id INTEGER PRIMARY KEY,
         date TEXT,
         meals TEXT,
-        water REAL
+        water INTEGER
       );
     ''');
   }
